+++ arc_application/forms/form.py
@@ -676,3 +676,157 @@ class OtherPeoplePreviousAddressManualForm(GOVUKForm):
         if re.match("^[A-Z]{1,2}[0-9]{1,2}[A-Z]?[0-9][A-Z][A-Z]$", postcode_uppercase) is None:
             raise forms.ValidationError('Please enter a valid postcode')
         return postcode
+
+
+class PersonalDetailsPreviousNames(GOVUKForm, ModelForm):
+    field_label_classes = 'form-label-bold'
+    auto_replace_widgets = True
+
+    def __init__(self, *args, **kwargs):
+        super(PersonalDetailsPreviousNames, self).__init__(*args, **kwargs)
+
+    class Meta:
+        model = PreviousName
+        fields = ['first_name', 'middle_names', 'last_name',
+                  'previous_name_id', 'person_id', 'other_person_type']
+        widgets = {
+            'previous_name_id': forms.HiddenInput(),
+            'person_id': forms.HiddenInput(),
+            'other_person_type': forms.HiddenInput()
+        }
+
+
+class PersonalDetailsPreviousPostcodeEntry(GOVUKForm):
+
+    field_label_classes = 'form-label-bold'
+    auto_replace_widgets = True
+
+    postcode = forms.CharField(label='Postcode', error_messages={'required': 'Please enter your postcode'})
+
+    def clean_postcode(self):
+        """
+        Postcode validation
+        :return: string
+        """
+        postcode = self.cleaned_data['postcode']
+        postcode_no_space = postcode.replace(" ", "")
+        postcode_uppercase = postcode_no_space.upper()
+        if re.match("^[A-Z]{1,2}[0-9]{1,2}[A-Z]?[0-9][A-Z][A-Z]$", postcode_uppercase) is None:
+            raise forms.ValidationError('Please enter a valid postcode')
+        return postcode
+
+
+class PersonalDetailsPreviousAddressLookupForm(GOVUKForm):
+
+    field_label_classes = 'form-label-bold'
+    auto_replace_widgets = True
+
+    address = forms.ChoiceField(label='Select address', required=True,
+                                error_messages={'required': 'Please select your address'})
+
+    def __init__(self, *args, **kwargs):
+        """
+        Method to configure the initialisation of the Your personal details: home address form for postcode search
+        :param args: arguments passed to the form
+        :param kwargs: keyword arguments passed to the form, e.g. application ID
+        """
+        self.choices = kwargs.pop('choices')
+        super(PersonalDetailsPreviousAddressLookupForm, self).__init__(*args, **kwargs)
+        self.fields['address'].choices = self.choices
+
+
+class PersonalDetailsPreviousAddressManualForm(GOVUKForm):
+    """
+    GOV.UK form for the Your personal details: home address page for manual entry
+    """
+    field_label_classes = 'form-label-bold'
+    auto_replace_widgets = True
+
+    street_name_and_number = forms.CharField(label='Address line 1', required=False)
+    street_name_and_number2 = forms.CharField(label='Address line 2', required=False)
+    town = forms.CharField(label='Town or city', required=False)
+    county = forms.CharField(label='County (optional)', required=False)
+    postcode = forms.CharField(label='Postcode', required=False)
+
+    def __init__(self, *args, **kwargs):
+        """
+        Method to configure the initialisation of the Your personal details: home address form for manual entry
+        :param args: arguments passed to the form
+        :param kwargs: keyword arguments passed to the form, e.g. application ID
+        """
+        try:
+            self.address_id = kwargs.pop('id')
+        except:
+            self.address_id = None
+
+        super(PersonalDetailsPreviousAddressManualForm, self).__init__(*args, **kwargs)
+        if PreviousAddress.objects.filter(previous_name_id=self.address_id).count() > 0:
+            previous_address = PreviousAddress.objects.get(previous_name_id=self.address_id)
+            self.fields['street_name_and_number'].initial = previous_address.street_line1
+            self.fields['street_name_and_number2'].initial = previous_address.street_line2
+            self.fields['town'].initial = previous_address.town
+            self.fields['county'].initial = previous_address.county
+            self.fields['postcode'].initial = previous_address.postcode
+            self.pk = previous_address.previous_name_id
+            self.field_list = ['street_name_and_number', 'street_name_and_number2', 'town', 'county', 'postcode']
+
+
+        # If information was previously entered, display it on the form
+
+
+    def clean_street_name_and_number(self):
+        """
+        Street name and number validation
+        :return: string
+        """
+        street_name_and_number = self.cleaned_data['street_name_and_number']
+        if len(street_name_and_number) > 50:
+            raise forms.ValidationError('The first line of your address must be under 50 characters long')
+        return street_name_and_number
+
+    def clean_street_name_and_number2(self):
+        """
+        Street name and number line 2 validation
+        :return: string
+        """
+        street_name_and_number2 = self.cleaned_data['street_name_and_number2']
+        if len(street_name_and_number2) > 50:
+            raise forms.ValidationError('The second line of your address must be under 50 characters long')
+        return street_name_and_number2
+
+    def clean_town(self):
+        """
+        Town validation
+        :return: string
+        """
+        town = self.cleaned_data['town']
+        if re.match("^[A-Za-z- ]+$", town) is None:
+            raise forms.ValidationError('Please spell out the name of the town or city using letters')
+        if len(town) > 50:
+            raise forms.ValidationError('The name of the town or city must be under 50 characters long')
+        return town
+
+    def clean_county(self):
+        """
+        County validation
+        :return: string
+        """
+        county = self.cleaned_data['county']
+        if county != '':
+            if re.match("^[A-Za-z- ]+$", county) is None:
+                raise forms.ValidationError('Please spell out the name of the county using letters')
+            if len(county) > 50:
+                raise forms.ValidationError('The name of the county must be under 50 characters long')
+        return county
+
+    def clean_postcode(self):
+        """
+        Postcode validation
+        :return: string
+        """
+        postcode = self.cleaned_data['postcode']
+        postcode_no_space = postcode.replace(" ", "")
+        postcode_uppercase = postcode_no_space.upper()
+        if re.match("^[A-Z]{1,2}[0-9]{1,2}[A-Z]?[0-9][A-Z][A-Z]$", postcode_uppercase) is None:
+            raise forms.ValidationError('Please enter a valid postcode')
+        return postcode
