+++ arc_application/templates/other-people-previous-address-manual-update.html
@@ -37,7 +37,7 @@
     <br>
     <!-- Return to task list -->
     <div class="form-group">
-        <a href="{{ list_url }}">Return to list</a>
+        <a href="{{ list_url }}" class="link-back">Return to list</a>
     </div>
 </form>
