+++ arc_application/templates/other-people-previous-address-select.html
@@ -69,7 +69,7 @@
     <br>
     <!-- Return to task list -->
     <div class="form-group">
-        <a href="{{ list_url }}">Return to list</a>
+        <a href="{{ list_url }}" class="link-back">Return to list</a>
     </div>
 </form>
